intel-iommu: dmar: Expand segment datatype to integer
The fake bus' domain number starts from 0x10000. But, the domain called
as segment in intel-iommu is a datatype u16, which has only 16 bits
width. That makes overflow and truncate to wrong value, then causes
unpredictable errors.

This patch expands segments datatype to integer to fix this issue.

https://phabricator.endlessm.com/T29119

+++ drivers/iommu/dmar.c
@@ -208,7 +208,7 @@ fallback:
 
 /* Return: > 0 if match found, 0 if no match found, < 0 if error happens */
 int dmar_insert_dev_scope(struct dmar_pci_notify_info *info,
-			  void *start, void*end, u16 segment,
+			  void *start, void*end, int segment,
 			  struct dmar_dev_scope *devices,
 			  int devices_cnt)
 {
@@ -264,7 +264,7 @@ int dmar_insert_dev_scope(struct dmar_pci_notify_info *info,
 	return 0;
 }
 
-int dmar_remove_dev_scope(struct dmar_pci_notify_info *info, u16 segment,
+int dmar_remove_dev_scope(struct dmar_pci_notify_info *info, int segment,
 			  struct dmar_dev_scope *devices, int count)
 {
 	int index;
